Add/edit recipe with ingredients completed

+++ recipe-app-data/src/main/java/gbc/comp3095/services/UserService.java
@@ -5,6 +5,9 @@ import gbc.comp3095.repositories.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
+import java.util.Set;
+
 @Service
 public class UserService {
     //*********************************************************************************
@@ -17,7 +20,6 @@ public class UserService {
 //* inherited in the User Repository interface
 // *********************************************************************************//
     private final UserRepository userRepository;
-    @Autowired
     public UserService(UserRepository userRepository) {
         this.userRepository = userRepository;
     }
@@ -28,25 +30,40 @@ public class UserService {
     }
 
     public User findById(Long id) {
-        return userRepository.findById(id).orElse(null);
+        List<User> saved_users = (List<User>) this.userRepository.findAll();
+        for (User u : saved_users) {
+            if (u.getId().equals(id))
+                return u;
+        }
+        return null;
     }
 
     public User save(User object) {
-        return userRepository.save(object);
+        return this.userRepository.save(object);
     }
 
     public void delete(User object) {
-        userRepository.delete(object);
+        if ((this.findById(object.getId()) != null) || (this.findByUsername(object.getUsername()) != null))
+            this.userRepository.delete(object);
     }
 
     public void deleteById(Long id) {
-        userRepository.deleteById(id);
+        User found_user = this.findById(id);
+        if (found_user != null)
+            userRepository.deleteById(id);
     }
 
     public Iterable<User> findAll() {
-        return userRepository.findAll();
+        return this.userRepository.findAll();
     }
 
-    public User findByUsername(String username) { return userRepository.findByUsername(username); }
+    public User findByUsername(String username) {
+        List<User> saved_users = (List<User>) this.userRepository.findAll();
+        for (User u : saved_users) {
+            if (u.getUsername().equals(username))
+                return u;
+        }
+        return null;
+    }
 
 }
